# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BAC MASTER CI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/goals:
    get:
      operationId: getDailyGoals
      tags: [users]
      summary: Get daily goals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Daily goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyGoals"
    put:
      operationId: updateDailyGoals
      tags: [users]
      summary: Update daily goals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDailyGoalsRequest"
      responses:
        "200":
          description: Updated goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyGoals"

  # SUBJECTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        series:
          type: string
          enum: [A, C, D]
        role:
          type: string
          enum: [student, admin]
        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - series
        - role
        - points
        - isPremium
        - createdAt
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        series:
          type: string
          enum: [A, C, D]
        avatarUrl:
          type: string
    DailyGoals:
      type: object
      properties:
        lessonsPerDay:
          type: integer
        exercisesPerDay:
          type: integer
        studyMinutesPerDay:
          type: integer
        lessonsCompletedToday:
          type: integer
        exercisesCompletedToday:
          type: integer
      required:
        - lessonsPerDay
        - exercisesPerDay
        - studyMinutesPerDay
        - lessonsCompletedToday
        - exercisesCompletedToday
    UpdateDailyGoalsRequest:
      type: object
      properties:
        lessonsPerDay:
          type: integer
        exercisesPerDay:
          type: integer
        studyMinutesPerDay:
          type: integer
...